requirements.txt
# MCP Server Dependencies
mcp>=1.0.0
fastmcp>=0.1.0
httpx>=0.27.0

# API Server Dependencies (FastAPI with Swagger UI)
fastapi>=0.115.0
uvicorn[standard]>=0.32.0

# Client Dependencies
langchain>=0.3.0
langchain-core>=0.3.0
langchain-openai>=0.2.0
langchain-mcp>=0.1.0
python-dotenv>=1.0.0

# Generic MCP Dependencies
pyyaml>=6.0.0
